+++ resources/views/bootstrap-5/components/buttons/form-button.blade.php
@@ -13,6 +13,9 @@
     @endif
     @if ($disabled === true) disabled @endif
 >
+    @if ($startContent !== null)
+        {!! $startContent !!}
+    @endif
     @if ($slot->isEmpty())
         @if ($hideText)
             <span class="visually-hidden">
@@ -23,6 +26,9 @@
         @endif
     @else
         {!! $slot !!}
+    @endif
+    @if ($endContent !== null)
+        {!! $endContent !!}
     @endif
 </button>