+++ src/Components/Buttons/Actions/Back/Back.php
@@ -22,6 +22,8 @@ class Back extends LinkButton
         public bool $disabled = false,
         public ?string $confirm = null,
         public ?string $confirmId = null,
+        public ?string $startContent = null,
+        public ?string $endContent = null,
     ) {
         $text ??= Str::ucfirst(trans('back.simple'));
 
@@ -42,6 +44,8 @@ class Back extends LinkButton
             disabled: $disabled,
             confirm: $confirm,
             confirmId: $confirmId,
+            startContent: $startContent,
+            endContent: $endContent,
         );
     }
 }